remove support from group_vars entirely
